refactor: seprate usePlaybackQueueController
