+++ src/App.tsx
@@ -1,14 +1,12 @@
 import React, { useState, useRef, useEffect, useCallback, useMemo } from 'react';
 import { useMotionValue } from 'framer-motion';
 import { useTranslation } from 'react-i18next';
-import { LyricParserFactory } from './utils/lyrics/LyricParserFactory';
-import { saveSessionData, getSessionData, getFromCache, getFromCacheWithMigration, saveToCache, getLocalSongs, removeFromCache, clearCacheByCategory } from './services/db';
+import { saveSessionData, getSessionData, getFromCache, saveToCache, removeFromCache, clearCacheByCategory } from './services/db';
 import { getCachedAudioBlob, hasCachedAudio, saveAudioBlob } from './services/audioCache';
-import { getCachedCoverUrl, loadCachedOrFetchCover } from './services/coverCache';
-import { ensureLocalSongEmbeddedCover, getAudioFromLocalSong } from './services/localMusicService';
-import { loadOnlineSongAudioSource, loadOnlineSongLyrics } from './services/onlinePlayback';
-import { buildLocalQueue, buildNavidromeQueue, buildUnifiedLocalSong, buildUnifiedNavidromeSong } from './services/playbackAdapters';
-import { getPrefetchedData, prefetchNearbySongs, invalidateAndRefetch, invalidatePrefetchedLyrics } from './services/prefetchService';
+import { loadCachedOrFetchCover } from './services/coverCache';
+import { loadOnlineSongAudioSource } from './services/onlinePlayback';
+import { buildNavidromeQueue } from './services/playbackAdapters';
+import { invalidatePrefetchedLyrics } from './services/prefetchService';
 import VisualizerRenderer from './components/visualizer/VisualizerRenderer';
 import AppShell from './components/app/AppShell';
 import Home from './components/app/Home';
@@ -19,44 +17,34 @@ import { useAppDialogsModel } from './components/app/view-models/useAppDialogsMo
 import { useAppOverlaysModel } from './components/app/view-models/useAppOverlaysModel';
 import { useHomeViewModel } from './components/app/view-models/useHomeViewModel';
 import { usePlayerPanelViewModel } from './components/app/view-models/usePlayerPanelViewModel';
-import { LyricData, Theme, PlayerState, SongResult, LocalSong, ReplayGainMode, LocalLibraryGroup, LocalPlaylist, UnifiedSong, StatusMessage, PlaybackContext, StageLoopMode } from './types';
+import { LyricData, Theme, PlayerState, SongResult, LocalSong, ReplayGainMode, LocalLibraryGroup, UnifiedSong, StatusMessage, PlaybackContext, StageLoopMode } from './types';
 import { NavidromeSong, NavidromeViewSelection } from './types/navidrome';
-import { NavidromeMatchData } from './components/modal/NaviLyricMatchModal';
-import { NextTrackOptions, PlaybackNavigationOptions, SkipPromptMessageKey, UnavailableReplacementRequest } from './types/appPlayback';
 import { getOnlineSongCacheKey, isCloudSong, isSongMarkedUnavailable, neteaseApi } from './services/netease';
 import { navidromeApi, getNavidromeConfig } from './services/navidromeService';
-import { addSongsToLocalPlaylist, createLocalPlaylist, getLocalPlaylists, removeSongsFromLocalPlaylist, setLocalSongFavorite } from './services/localPlaylistService';
+import { removeSongsFromLocalPlaylist } from './services/localPlaylistService';
 import { useAppNavigation } from './hooks/useAppNavigation';
 import { useNeteaseLibrary } from './hooks/useNeteaseLibrary';
 import { useAppPreferences } from './hooks/useAppPreferences';
 import { useElectronPlaybackBridge } from './hooks/useElectronPlaybackBridge';
 import { useMediaSessionBridge } from './hooks/useMediaSessionBridge';
 import { usePlaybackUiEffects } from './hooks/usePlaybackUiEffects';
+import { useLibraryPlaybackController } from './hooks/useLibraryPlaybackController';
+import { usePlaybackQueueController } from './hooks/usePlaybackQueueController';
 import { useSessionRestoreController } from './hooks/useSessionRestoreController';
 import { useStagePlaybackController } from './hooks/useStagePlaybackController';
 import { useThemeController } from './hooks/useThemeController';
 import { useSearchNavigationStore } from './stores/useSearchNavigationStore';
 import { useShallow } from 'zustand/react/shallow';
-import { hydrateNavidromeLyricPayload, resolvePreferredNavidromeLyrics } from './utils/appNavidromeLyrics';
-import { clampMediaVolume, extractCloudLyricText, findLatestActiveLineIndex, formatTime, getAudioSrcKind, hasRenderableLyrics, replayGainModeLabels, resolveDebugLyricsSource, resolveDebugSongSource, toDebugLineSnapshot, toSafeRemoteUrl } from './utils/appPlaybackHelpers';
-import { isLocalPlaybackSong, isNavidromePlaybackSong, isStagePlaybackSong, resolveNavidromePlaybackCarrier } from './utils/appPlaybackGuards';
+import { clampMediaVolume, findLatestActiveLineIndex, formatTime, getAudioSrcKind, replayGainModeLabels, resolveDebugLyricsSource, resolveDebugSongSource, toDebugLineSnapshot, toSafeRemoteUrl } from './utils/appPlaybackHelpers';
+import { isLocalPlaybackSong, isNavidromePlaybackSong, isStagePlaybackSong } from './utils/appPlaybackGuards';
 import { getNextLoopMode, getStageLyricsTimelineBounds } from './utils/appStageHelpers';
-import { isPureMusicLyricText } from './utils/lyrics/pureMusic';
-import { ensureLyricDataRenderHints, getLineRenderHints, migrateLyricDataRenderHints } from './utils/lyrics/renderHints';
-import { migrateMatchedLyricsCarrierRenderHints } from './utils/lyrics/storageMigration';
-import { processNeteaseLyrics } from './utils/lyrics/neteaseProcessing';
+import { ensureLyricDataRenderHints, getLineRenderHints } from './utils/lyrics/renderHints';
 import { applyLyricDisplayFilter } from './utils/lyrics/filtering';
 
 const LOCAL_MUSIC_UPDATED_EVENT = 'folia-local-music-updated';
-const LOCAL_PREWARM_OFFSETS = [-1, 1, 2] as const;
-const LOCAL_PREWARM_DELAY_MS = 1000;
 const DEV_DEBUG_SHORTCUT_LABEL = 'Alt+Shift+D';
 const ONLINE_AUDIO_URL_TTL_MS = 1200 * 1000;
 const ONLINE_AUDIO_URL_REFRESH_BUFFER_MS = 60 * 1000;
-const MAX_UNAVAILABLE_AUTO_SKIP_COUNT = 2;
-const UNAVAILABLE_SKIP_CONFIRM_TIMEOUT_MS = 5000;
-const UNAVAILABLE_SKIP_CONFIRM_INTERVAL_MS = 1000;
-
 // Default Theme
 // 午夜墨染
 const DEFAULT_THEME: Theme = {
@@ -150,12 +138,8 @@ export default function App() {
     const lastAudioRecoverySourceRef = useRef<string | null>(null);
     const currentOnlineAudioUrlFetchedAtRef = useRef<number | null>(null);
     const [isLyricsLoading, setIsLyricsLoading] = useState(false);
-    const [pendingUnavailableReplacement, setPendingUnavailableReplacement] = useState<UnavailableReplacementRequest | null>(null);
     const isNowPlayingControlDisabledRef = useRef(false);
 
-    // Local Music State
-    const [localSongs, setLocalSongs] = useState<LocalSong[]>([]);
-    const [localPlaylists, setLocalPlaylists] = useState<LocalPlaylist[]>([]);
     const [replayGainMode, setReplayGainMode] = useState<ReplayGainMode>(() => {
         const saved = localStorage.getItem('local_replaygain_mode');
         return saved === 'track' || saved === 'album' ? saved : 'off';
@@ -666,28 +650,63 @@ export default function App() {
 
     // --- Local Music Functions ---
 
-    const loadLocalSongs = async () => {
-        try {
-            const songs = await getLocalSongs();
-            setLocalSongs(songs);
-        } catch (error) {
-            console.error('Failed to load local songs:', error);
-        }
-    };
-
-    const loadLocalPlaylists = useCallback(async () => {
-        try {
-            const playlists = await getLocalPlaylists();
-            setLocalPlaylists(playlists);
-        } catch (error) {
-            console.error('Failed to load local playlists:', error);
-        }
-    }, []);
-
-    const onRefreshLocalSongs = async () => {
-        await loadLocalSongs();
-        await loadLocalPlaylists();
-    };
+    const {
+        localSongs,
+        localPlaylists,
+        showLyricMatchModal,
+        setShowLyricMatchModal,
+        showNaviLyricMatchModal,
+        setShowNaviLyricMatchModal,
+        loadLocalSongs,
+        loadLocalPlaylists,
+        onRefreshLocalSongs,
+        isLocalSongLiked,
+        saveCurrentQueueAsLocalPlaylist,
+        addCurrentSongToLocalPlaylist,
+        createCurrentLocalPlaylist,
+        addCurrentSongToNeteasePlaylist,
+        addCurrentSongToNavidromePlaylist,
+        createCurrentNavidromePlaylist,
+        loadCurrentSongLyricPreview,
+        handleLocalQueueAdd,
+        onPlayLocalSong,
+        onPlayNavidromeSong,
+        onMatchNavidromeSong,
+        handleUpdateLocalLyrics,
+        handleChangeLyricsSource,
+        handleManualMatchOnline,
+        handleLyricMatchComplete,
+        handleNaviLyricMatchComplete,
+        handleHomeMatchSong,
+        handleLike: handleLibraryLike,
+    } = useLibraryPlaybackController({
+        t: (key, fallback) => t(key, fallback ?? ''),
+        audioQuality,
+        currentSong,
+        lyrics,
+        playQueue,
+        userId: user?.userId,
+        currentTime,
+        setCurrentSong,
+        setLyrics,
+        setCachedCoverUrl,
+        setAudioSrc,
+        setPlayQueue,
+        setPlayerState,
+        setCurrentLineIndex,
+        setDuration,
+        setIsLyricsLoading,
+        setStatusMsg,
+        setIsPanelOpen,
+        navigateToPlayer,
+        persistLastPlaybackCache,
+        restoreCachedThemeForSong,
+        interruptStagePlaybackForMainTransition,
+        blobUrlRef,
+        shouldAutoPlayRef: shouldAutoPlay,
+        currentSongRef,
+        currentOnlineAudioUrlFetchedAtRef,
+    });
 
     useSessionRestoreController({
         audioQuality,
@@ -706,19 +725,6 @@ export default function App() {
         loadLocalSongs,
         loadLocalPlaylists,
     });
-
-    const getFavoriteLocalPlaylist = useMemo(
-        () => localPlaylists.find(playlist => playlist.isFavorite) ?? null,
-        [localPlaylists]
-    );
-
-    const isLocalSongLiked = useCallback((song: SongResult | null) => {
-        if (!song || !isLocalPlaybackSong(song) || !song.localData || !getFavoriteLocalPlaylist) {
-            return false;
-        }
-
-        return getFavoriteLocalPlaylist.songIds.includes(song.localData.id);
-    }, [getFavoriteLocalPlaylist]);
 
     const openLocalLibraryGroup = useCallback((group: LocalLibraryGroup, row: 0 | 1 | 2 | 3) => {
         setHomeViewTab('local');
@@ -847,285 +853,6 @@ export default function App() {
         }, 2);
     }, [localSongs, openLocalLibraryGroup]);
 
-    const saveCurrentQueueAsLocalPlaylist = useCallback(async (name: string) => {
-        const trimmedName = name.trim();
-        if (!trimmedName) {
-            throw new Error('Playlist name is empty');
-        }
-
-        const queueSongs = playQueue
-            .map(song => (song as any).localData as LocalSong | undefined)
-            .filter((song): song is LocalSong => Boolean(song?.id));
-
-        if (!queueSongs.length) {
-            throw new Error('No local songs in queue');
-        }
-
-        await createLocalPlaylist(trimmedName, queueSongs);
-        await loadLocalPlaylists();
-    }, [loadLocalPlaylists, playQueue]);
-
-    const addCurrentSongToLocalPlaylist = useCallback(async (playlistId: string) => {
-        if (!isLocalPlaybackSong(currentSong) || !currentSong.localData) {
-            throw new Error('Current song is not local');
-        }
-
-        await addSongsToLocalPlaylist(playlistId, [currentSong.localData]);
-        await loadLocalPlaylists();
-    }, [currentSong, loadLocalPlaylists]);
-
-    const createCurrentLocalPlaylist = useCallback(async (name: string) => {
-        const trimmedName = name.trim();
-        if (!trimmedName) {
-            throw new Error('Playlist name is empty');
-        }
-
-        if (!isLocalPlaybackSong(currentSong) || !currentSong.localData) {
-            throw new Error('Current song is not local');
-        }
-
-        await createLocalPlaylist(trimmedName, [currentSong.localData]);
-        await loadLocalPlaylists();
-        setStatusMsg({ type: 'success', text: t('status.playlistUpdated') || '歌单已更新' });
-    }, [currentSong, loadLocalPlaylists, t]);
-
-    const addCurrentSongToNeteasePlaylist = useCallback(async (playlistId: number) => {
-        if (!currentSong || isLocalPlaybackSong(currentSong) || isNavidromePlaybackSong(currentSong)) {
-            throw new Error('Current song is not a Netease song');
-        }
-
-        await neteaseApi.updatePlaylistTracks('add', playlistId, [currentSong.id]);
-        await removeFromCache(`playlist_tracks_${playlistId}`);
-        await removeFromCache(`playlist_detail_${playlistId}`);
-        await refreshUserData();
-        setStatusMsg({ type: 'success', text: t('status.playlistUpdated') || '歌单已更新' });
-    }, [currentSong, refreshUserData, t]);
-
-    const addCurrentSongToNavidromePlaylist = useCallback(async (playlistId: string) => {
-        if (!isNavidromePlaybackSong(currentSong)) {
-            throw new Error('Current song is not a Navidrome song');
-        }
-
-        const config = getNavidromeConfig();
-        const navidromeSong = resolveNavidromePlaybackCarrier(currentSong);
-        if (!config || !navidromeSong?.navidromeData?.id) {
-            throw new Error('Navidrome is not configured');
-        }
-
-        await navidromeApi.updatePlaylist(config, playlistId, {
-            songIdsToAdd: [navidromeSong.navidromeData.id],
-        });
-        setStatusMsg({ type: 'success', text: t('status.playlistUpdated') || '歌单已更新' });
-    }, [currentSong, t]);
-
-    const createCurrentNavidromePlaylist = useCallback(async (name: string) => {
-        if (!isNavidromePlaybackSong(currentSong)) {
-            throw new Error('Current song is not a Navidrome song');
-        }
-
-        const config = getNavidromeConfig();
-        const navidromeSong = resolveNavidromePlaybackCarrier(currentSong);
-        if (!config || !navidromeSong?.navidromeData?.id) {
-            throw new Error('Navidrome is not configured');
-        }
-
-        await navidromeApi.createPlaylist(config, name, [navidromeSong.navidromeData.id]);
-        setStatusMsg({ type: 'success', text: t('status.playlistUpdated') || '歌单已更新' });
-    }, [currentSong, t]);
-
-    const handleLocalSongMatch = async (localSong: LocalSong): Promise<{ updatedLocalSong: LocalSong, matchedSongResult: SongResult | null; }> => {
-        let updatedLocalSong = localSong;
-        let matchedSongResult: SongResult | null = null;
-
-        // Only match online if: no local lyrics AND (no matched lyrics OR no matched cover) AND auto-match is enabled
-        const needsLyricsMatch = !localSong.hasLocalLyrics && !localSong.hasEmbeddedLyrics && !localSong.matchedLyrics;
-        const needsCoverMatch = !localSong.embeddedCover && !localSong.matchedCoverUrl;
-        if ((needsLyricsMatch || needsCoverMatch) && !localSong.noAutoMatch) {
-            setStatusMsg({ type: 'info', text: '正在匹配歌词和封面...' });
-            try {
-                const { matchLyrics } = await import('./services/localMusicService');
-                await matchLyrics(localSong);
-
-                // Reload local song to pick up cover-only or metadata-only matches as well.
-                const updatedSongs = await getLocalSongs();
-                const found = updatedSongs.find(s => s.id === localSong.id);
-
-                if (found) {
-                    updatedLocalSong = found;
-
-                    // Get full matched song details for UI
-                    if (found.matchedSongId) {
-                        try {
-                            const searchRes = await neteaseApi.cloudSearch(
-                                localSong.artist
-                                    ? `${localSong.artist} ${localSong.title}`
-                                    : localSong.title || localSong.fileName
-                            );
-                            if (searchRes.result?.songs) {
-                                matchedSongResult = searchRes.result.songs.find(s => s.id === found.matchedSongId) || searchRes.result.songs[0];
-                            }
-                        } catch (e) {
-                            console.warn('Failed to get matched song details:', e);
-                        }
-                    }
-                }
-            } catch (error) {
-                console.warn('Auto-match failed:', error);
-            }
-            // Refresh local songs list in App state
-            await loadLocalSongs();
-        }
-
-        return { updatedLocalSong, matchedSongResult };
-    };
-
-    const resolveLocalMetadataUI = async (localData: LocalSong, matchedSong: SongResult | null) => {
-        let embeddedCoverUrl: string | null = null;
-        if (localData.embeddedCover) {
-            embeddedCoverUrl = URL.createObjectURL(localData.embeddedCover);
-        }
-
-        const preferOnlineCover = localData.useOnlineCover === true;
-        const preferOnlineMetadata = localData.useOnlineMetadata === true;
-
-        const coverUrl = preferOnlineCover
-            ? (localData.matchedCoverUrl || embeddedCoverUrl || null)
-            : (embeddedCoverUrl || localData.matchedCoverUrl || null);
-
-        let lyrics: LyricData | null = null;
-        const source = localData.lyricsSource;
-        if (source === 'online' && localData.matchedLyrics) {
-            lyrics = localData.matchedLyrics;
-        } else if (source === 'embedded' && localData.embeddedLyricsContent) {
-            lyrics = await LyricParserFactory.parse({ type: 'embedded', textContent: localData.embeddedLyricsContent, translationContent: localData.embeddedTranslationLyricsContent });
-        } else if (source === 'local' && localData.localLyricsContent) {
-            lyrics = await LyricParserFactory.parse({ type: 'local', lrcContent: localData.localLyricsContent, tLrcContent: localData.localTranslationLyricsContent });
-        } else if (!source) {
-            if (localData.hasLocalLyrics && localData.localLyricsContent) {
-                lyrics = await LyricParserFactory.parse({ type: 'local', lrcContent: localData.localLyricsContent, tLrcContent: localData.localTranslationLyricsContent });
-            } else if (localData.hasEmbeddedLyrics && localData.embeddedLyricsContent) {
-                lyrics = await LyricParserFactory.parse({ type: 'embedded', textContent: localData.embeddedLyricsContent, translationContent: localData.embeddedTranslationLyricsContent });
-            } else if (localData.matchedLyrics) {
-                lyrics = localData.matchedLyrics;
-            }
-        }
-
-        const unifiedSong = buildUnifiedLocalSong({
-            localSong: localData,
-            matchedSong,
-            coverUrl,
-            preferOnlineMetadata,
-        });
-
-        return { lyrics, coverUrl, unifiedSong };
-    };
-
-    const loadCurrentSongLyricPreview = useCallback(async (): Promise<LyricData | null> => {
-        if (!currentSong) {
-            return null;
-        }
-
-        if (isLocalPlaybackSong(currentSong) && currentSong.localData) {
-            const localData = currentSong.localData;
-            const source = localData.lyricsSource;
-
-            if (source === 'online' && localData.matchedLyrics) {
-                return localData.matchedLyrics;
-            }
-            if (source === 'embedded' && localData.embeddedLyricsContent) {
-                return await LyricParserFactory.parse({
-                    type: 'embedded',
-                    textContent: localData.embeddedLyricsContent,
-                    translationContent: localData.embeddedTranslationLyricsContent,
-                });
-            }
-            if (source === 'local' && localData.localLyricsContent) {
-                return await LyricParserFactory.parse({
-                    type: 'local',
-                    lrcContent: localData.localLyricsContent,
-                    tLrcContent: localData.localTranslationLyricsContent,
-                });
-            }
-            if (!source) {
-                if (localData.hasLocalLyrics && localData.localLyricsContent) {
-                    return await LyricParserFactory.parse({
-                        type: 'local',
-                        lrcContent: localData.localLyricsContent,
-                        tLrcContent: localData.localTranslationLyricsContent,
-                    });
-                }
-                if (localData.hasEmbeddedLyrics && localData.embeddedLyricsContent) {
-                    return await LyricParserFactory.parse({
-                        type: 'embedded',
-                        textContent: localData.embeddedLyricsContent,
-                        translationContent: localData.embeddedTranslationLyricsContent,
-                    });
-                }
-                if (localData.matchedLyrics) {
-                    return localData.matchedLyrics;
-                }
-            }
-
-            return lyrics;
-        }
-
-        if (isNavidromePlaybackSong(currentSong)) {
-            const navidromeSong = resolveNavidromePlaybackCarrier(currentSong);
-            if (!navidromeSong) {
-                return lyrics;
-            }
-
-            if ((navidromeSong as any).lyricsSource === 'online' && (navidromeSong as any).matchedLyrics) {
-                return (navidromeSong as any).matchedLyrics as LyricData;
-            }
-
-            let resolved = await resolvePreferredNavidromeLyrics(navidromeSong);
-            if (resolved) {
-                return resolved;
-            }
-
-            const config = getNavidromeConfig();
-            if (config) {
-                await hydrateNavidromeLyricPayload(config, navidromeSong);
-                resolved = await resolvePreferredNavidromeLyrics(navidromeSong);
-                if (resolved) {
-                    return resolved;
-                }
-            }
-
-            return lyrics;
-        }
-
-        const onlineSong = currentSong;
-        const cachedLyrics = await getFromCacheWithMigration<LyricData>(
-            getOnlineSongCacheKey('lyric', onlineSong),
-            migrateLyricDataRenderHints
-        );
-        if (cachedLyrics) {
-            return cachedLyrics;
-        }
-
-        const prefetched = getPrefetchedData(onlineSong, audioQuality);
-        if (prefetched?.lyrics) {
-            return prefetched.lyrics;
-        }
-
-        if (isCloudSong(onlineSong) && user?.userId) {
-            const lyricRes = await neteaseApi.getCloudLyric(user.userId, onlineSong.id);
-            const mainLrc = extractCloudLyricText(lyricRes);
-            if (!mainLrc || isPureMusicLyricText(mainLrc)) {
-                return null;
-            }
-
-            return await LyricParserFactory.parse({
-                type: 'local',
-                lrcContent: mainLrc,
-            });
-        }
-
-        const lyricRes = await neteaseApi.getLyric(onlineSong.id);
-        return (await processNeteaseLyrics(neteaseApi.getProcessedLyricPayload(lyricRes))).lyrics;
-    }, [audioQuality, currentSong, lyrics, user?.userId]);
 
     const handleSaveLyricFilterPattern = useCallback(async (pattern: string) => {
         handleSetLyricFilterPattern(pattern);
@@ -1179,17 +906,6 @@ export default function App() {
         return false;
     }, [activePlaybackContext, currentSong, persistLastPlaybackCache, playQueue, t]);
 
-    const handleLocalQueueAdd = async (localSong: LocalSong) => {
-        const preparedLocalSong = await ensureLocalSongEmbeddedCover(localSong);
-        const { unifiedSong } = await resolveLocalMetadataUI(preparedLocalSong, null);
-        const exists = playQueue.some(song => song.id === unifiedSong.id);
-        const nextQueue = exists ? playQueue : [...playQueue, unifiedSong];
-
-        setPlayQueue(nextQueue);
-        void persistLastPlaybackCache(currentSong, nextQueue);
-        setStatusMsg({ type: 'success', text: t('status.queueUpdated') || '已添加到播放队列' });
-    };
-
     const addNeteaseSongToQueue = useCallback((song: SongResult) => {
         if (isSongMarkedUnavailable(song)) {
             return;
@@ -1217,303 +933,84 @@ export default function App() {
         setStatusMsg({ type: 'success', text: t('status.queueUpdated') || '已添加到播放队列' });
     }, [currentSong, persistLastPlaybackCache, playQueue, t]);
 
-    const prewarmLocalSongMetadata = async (localSong: LocalSong) => {
-        const preparedLocalSong = await ensureLocalSongEmbeddedCover(localSong);
-        Object.assign(localSong, preparedLocalSong);
-
-        const needsLyricsMatch = !localSong.hasLocalLyrics && !localSong.hasEmbeddedLyrics && !localSong.matchedLyrics;
-        const needsCoverMatch = !localSong.embeddedCover && !localSong.matchedCoverUrl;
-
-        if ((needsLyricsMatch || needsCoverMatch) && !localSong.noAutoMatch) {
-            try {
-                const { matchLyrics } = await import('./services/localMusicService');
-                await matchLyrics(localSong);
-            } catch (error) {
-                console.warn('[LocalPrewarm] Failed to prewarm local song metadata:', error);
-            }
-        }
-    };
-
-    const prewarmNearbyLocalSongs = (currentSong: LocalSong, queue: LocalSong[] = []) => {
-        if (queue.length === 0) {
-            return;
-        }
-
-        const currentIndex = queue.findIndex(song => song.id === currentSong.id);
-        if (currentIndex === -1) {
-            return;
-        }
-
-        const nearbySongs = LOCAL_PREWARM_OFFSETS
-            .map(offset => queue[currentIndex + offset])
-            .filter((song): song is LocalSong => Boolean(song));
-
-        if (nearbySongs.length === 0) {
-            return;
-        }
-
-        window.setTimeout(() => {
-            void (async () => {
-                for (const nearbySong of nearbySongs) {
-                    await prewarmLocalSongMetadata(nearbySong);
-                }
-            })();
-        }, LOCAL_PREWARM_DELAY_MS);
-    };
-
-    const onPlayLocalSong = async (localSong: LocalSong, queue: LocalSong[] = []) => {
-        interruptStagePlaybackForMainTransition();
-
-        // Get audio blob from fileHandle first
-        const blobUrl = await getAudioFromLocalSong(localSong);
-        if (!blobUrl) {
-            setStatusMsg({
-                type: 'error',
-                text: '无法访问文件，请重新导入文件夹'
-            });
-            return;
-        }
-
-        const preparedLocalSong = await ensureLocalSongEmbeddedCover(localSong);
-        const initialMeta = await resolveLocalMetadataUI(preparedLocalSong, null);
-
-        if (blobUrlRef.current) URL.revokeObjectURL(blobUrlRef.current);
-        blobUrlRef.current = blobUrl;
-
-        shouldAutoPlay.current = true;
-        currentSongRef.current = initialMeta.unifiedSong.id;
-
-        setLyrics(initialMeta.lyrics);
-        setCurrentLineIndex(-1);
-        currentTime.set(0);
-        setCurrentSong(initialMeta.unifiedSong);
-        setAudioSrc(blobUrl);
-
-        if (initialMeta.coverUrl) {
-            loadCachedOrFetchCover(`cover_local_${preparedLocalSong.id}`, initialMeta.coverUrl).then(res => {
-                if (currentSongRef.current === initialMeta.unifiedSong.id) setCachedCoverUrl(res);
-            });
-        } else {
-            setCachedCoverUrl(null);
-        }
-
-        setIsLyricsLoading(true); // Assuming we might fetch more metadata online
-
-        // Set queue
-        if (queue.length > 0) {
-            const finalQueue = buildLocalQueue(queue, initialMeta.unifiedSong);
-            setPlayQueue(finalQueue);
-            void persistLastPlaybackCache(initialMeta.unifiedSong, finalQueue);
-        } else {
-            setPlayQueue([initialMeta.unifiedSong]);
-            void persistLastPlaybackCache(initialMeta.unifiedSong, [initialMeta.unifiedSong]);
-        }
-        navigateToPlayer();
-        setPlayerState(PlayerState.IDLE);
-        setStatusMsg({ type: 'success', text: '本地音乐已加载' });
-        prewarmNearbyLocalSongs(preparedLocalSong, queue);
-
-        // --- Background Auto-Match ---
-        handleLocalSongMatch(preparedLocalSong).then(async ({ updatedLocalSong, matchedSongResult }) => {
-            if (currentSongRef.current !== initialMeta.unifiedSong.id) return; // User skipped track
-            
-            const updatedMeta = await resolveLocalMetadataUI(updatedLocalSong, matchedSongResult);
-            
-            setCurrentSong(updatedMeta.unifiedSong);
-            setLyrics(updatedMeta.lyrics);
-            setIsLyricsLoading(false);
-
-            if (updatedMeta.coverUrl && updatedMeta.coverUrl !== initialMeta.coverUrl) {
-                loadCachedOrFetchCover(`cover_local_${updatedLocalSong.id}`, updatedMeta.coverUrl).then(res => {
-                    if (currentSongRef.current === updatedMeta.unifiedSong.id) setCachedCoverUrl(res);
-                });
-            } else if (!updatedMeta.coverUrl) {
-                setCachedCoverUrl(null);
-            }
-        });
-    };
-
-    // --- Navidrome Playback ---
-    const onPlayNavidromeSong = async (
-        navidromeSong: NavidromeSong,
-        queue: NavidromeSong[] = [],
-        options: PlaybackNavigationOptions = {}
-    ) => {
-        interruptStagePlaybackForMainTransition();
-
-        const shouldNavigateToPlayer = options.shouldNavigateToPlayer ?? true;
-        const config = getNavidromeConfig();
-        if (!config) {
-            setStatusMsg({ type: 'error', text: 'Navidrome not configured' });
-            return;
-        }
-
-        setIsLyricsLoading(true);
-
-        try {
-            // Get streaming URL using navidromeData.id
-            const navidromeId = navidromeSong.navidromeData.id;
-            const streamUrl = navidromeApi.getStreamUrl(config, navidromeId);
-
-            // Fetch match data if available
-            const matchData = await getFromCacheWithMigration<NavidromeMatchData>(
-                `navidrome_match_${navidromeId}`,
-                migrateMatchedLyricsCarrierRenderHints
-            );
-
-            let lyrics: LyricData | null = null;
-            let coverUrl: string | undefined;
-            let showedLoadingToast = false;
-
-            if (matchData) {
-                if (matchData.lyricsSource === 'online' && matchData.matchedLyrics) {
-                    lyrics = matchData.matchedLyrics;
-                    console.log('[App] Using manually matched OpenSubsonic online lyrics');
-                }
-                if (matchData.useOnlineCover && matchData.matchedCoverUrl) {
-                    coverUrl = matchData.matchedCoverUrl;
-                }
-            }
-
-            if (!lyrics) {
-                lyrics = await resolvePreferredNavidromeLyrics(navidromeSong);
-                if (hasRenderableLyrics(lyrics)) {
-                    console.log('[App] Using cached Navidrome lyrics');
-                }
-            }
-
-            if (!lyrics) {
-                if (!showedLoadingToast) {
-                    setStatusMsg({ type: 'info', text: t('status.loadingSong') || '加载歌曲中...' });
-                    showedLoadingToast = true;
-                }
-                await hydrateNavidromeLyricPayload(config, navidromeSong);
-                lyrics = await resolvePreferredNavidromeLyrics(navidromeSong);
-                if (hasRenderableLyrics(lyrics)) {
-                    console.log('[App] Using embedded Navidrome lyrics');
-                }
-            }
-
-            // If no lyrics from Navidrome, try Netease (Auto Match)
-            let isAutoMatched = false;
-            let autoMatchedLyrics: LyricData | null = null;
-            if (!lyrics && !matchData?.noAutoMatch) {
-                try {
-                    if (!showedLoadingToast) {
-                        setStatusMsg({ type: 'info', text: t('status.loadingSong') || '加载歌曲中...' });
-                        showedLoadingToast = true;
-                    }
-                    const artistName = navidromeSong.artists?.[0]?.name || navidromeSong.ar?.[0]?.name || '';
-                    const searchQuery = `${navidromeSong.name} ${artistName}`.trim();
-                    const searchRes = await neteaseApi.cloudSearch(searchQuery, 1);
-
-                    if (searchRes.result?.songs?.length > 0) {
-                        const matchedSong = searchRes.result.songs[0];
-                        const lyricRes = await neteaseApi.getLyric(matchedSong.id);
-                        const processed = await processNeteaseLyrics({
-                            type: 'netease',
-                            ...lyricRes
-                        });
-
-                        lyrics = processed.lyrics;
-                        (navidromeSong as any).matchedIsPureMusic = processed.isPureMusic;
-                        
-                        if (lyrics || processed.isPureMusic) {
-                            autoMatchedLyrics = lyrics;
-                            isAutoMatched = true;
-                            console.log('[App] Using Netease lyrics for Navidrome song');
-                        }
-                    }
-                } catch (e) {
-                    console.warn('[App] Failed to fetch Netease lyrics for Navidrome song:', e);
-                }
-            }
-
-            // Attach match properties for NaviTab logic
-            if (isAutoMatched) {
-                (navidromeSong as any).matchedLyrics = autoMatchedLyrics;
-                (navidromeSong as any).useOnlineLyrics = true;
-                (navidromeSong as any).lyricsSource = 'online';
-            } else {
-                (navidromeSong as any).matchedLyrics = matchData?.matchedLyrics;
-                (navidromeSong as any).matchedIsPureMusic = matchData?.matchedIsPureMusic;
-                (navidromeSong as any).useOnlineLyrics = matchData?.useOnlineLyrics;
-                (navidromeSong as any).lyricsSource =
-                    matchData?.lyricsSource === 'online'
-                        ? 'online'
-                        : (hasRenderableLyrics(lyrics) ? 'navi' : matchData?.lyricsSource);
-            }
-
-            // Get cover art URL
-            if (!coverUrl) {
-                coverUrl = navidromeSong.album?.picUrl || navidromeSong.al?.picUrl ||
-                    navidromeApi.getCoverArtUrl(config, navidromeId);
-            }
-
-            // Create unified song for playback
-            const unifiedSong = buildUnifiedNavidromeSong(navidromeSong, {
-                coverUrl,
-                useOnlineMetadata: matchData?.useOnlineMetadata,
-                matchedArtists: matchData?.matchedArtists,
-                matchedAlbumName: matchData?.matchedAlbumName,
-            });
-
-            // Enable autoplay
-            shouldAutoPlay.current = true;
-            currentSongRef.current = unifiedSong.id;
-
-            // Set UI state
-            setLyrics(lyrics);
-            setCurrentLineIndex(-1);
-            currentTime.set(0);
-            setCurrentSong(unifiedSong);
-            setCachedCoverUrl(coverUrl);
-            setAudioSrc(streamUrl);
-            setIsLyricsLoading(false);
-
-            // Set queue
-            if (queue.length > 0) {
-                const finalQueue = buildNavidromeQueue(queue, unifiedSong);
-                setPlayQueue(finalQueue);
-                void persistLastPlaybackCache(unifiedSong, finalQueue);
-            } else {
-                setPlayQueue([unifiedSong]);
-                void persistLastPlaybackCache(unifiedSong, [unifiedSong]);
-            }
-
-            if (shouldNavigateToPlayer) {
-                navigateToPlayer();
-            }
-            setPlayerState(PlayerState.IDLE);
-            setStatusMsg({ type: 'success', text: 'Navidrome 歌曲已加载' });
-        } catch (e) {
-            console.error('[App] Failed to play Navidrome song:', e);
-            setStatusMsg({ type: 'error', text: '播放失败' });
-            setIsLyricsLoading(false);
-        }
-    };
-
-    // --- Navidrome Lyric Matching ---
-    const onMatchNavidromeSong = async (navidromeSong: NavidromeSong) => {
-        // This opens the lyric matching modal for Navidrome songs
-        // For now, we'll use a similar approach to local song matching
-        // The actual matching will be handled when playing the song
-        setStatusMsg({ type: 'info', text: t('navidrome.fetchingLyrics') || '正在匹配歌词...' });
-    };
-
     // --- Effects ---
 
-    const clearPendingUnavailableSkip = useCallback(() => {
-        if (pendingUnavailableSkipTimerRef.current !== null) {
-            window.clearTimeout(pendingUnavailableSkipTimerRef.current);
-            pendingUnavailableSkipTimerRef.current = null;
-        }
-
-        if (pendingUnavailableSkipIntervalRef.current !== null) {
-            window.clearInterval(pendingUnavailableSkipIntervalRef.current);
-            pendingUnavailableSkipIntervalRef.current = null;
-        }
-    }, []);
+    const {
+        pendingUnavailableReplacement,
+        setPendingUnavailableReplacement,
+        clearPendingUnavailableSkip,
+        playSong,
+        playOnlineQueueFromStart,
+        handleQueueAddAndPlay,
+        handleSearchOverlaySubmit,
+        handleSearchLoadMore,
+        handleSearchResultPlay,
+        handleUnavailableReplacementConfirm,
+        handleSearchResultArtistSelect,
+        handleSearchResultAlbumSelect,
+        handleNextTrack,
+        handlePrevTrack,
+        skipAfterPlaybackFailure,
+        handleStageExternalPlayRequest,
+        shuffleQueue,
+    } = usePlaybackQueueController({
+        t,
+        audioQuality,
+        currentSong,
+        playQueue,
+        playerState,
+        loopMode,
+        isFmMode,
+        isNowPlayingStageActive,
+        searchQuery,
+        searchSourceTab,
+        localSongs,
+        userId: user?.userId,
+        currentTime,
+        setCurrentSong,
+        setLyrics,
+        setCachedCoverUrl,
+        setAudioSrc,
+        setPlayQueue,
+        setPlayerState,
+        setCurrentLineIndex,
+        setDuration,
+        setIsLyricsLoading,
+        setStatusMsg,
+        setIsFmMode,
+        setPanelTab,
+        setIsPanelOpen,
+        navigateToPlayer,
+        navigateToSearch,
+        hideSearchOverlay,
+        setHomeViewTab,
+        setPendingNavidromeSelection,
+        handleArtistSelect,
+        handleAlbumSelect,
+        openLocalArtistByName,
+        openLocalAlbumByName,
+        appendNeteaseSongsToMainQueue,
+        persistLastPlaybackCache,
+        restoreCachedThemeForSong,
+        interruptStagePlaybackForMainTransition,
+        onPlayLocalSong,
+        onPlayNavidromeSong,
+        searchDeps: {
+            submitSearch,
+            loadMoreSearchResults,
+        },
+        audioRef,
+        blobUrlRef,
+        shouldAutoPlayRef: shouldAutoPlay,
+        currentSongRef,
+        playbackRequestIdRef,
+        playbackAutoSkipCountRef,
+        pendingUnavailableSkipTimerRef,
+        pendingUnavailableSkipIntervalRef,
+        pendingResumeTimeRef,
+        currentOnlineAudioUrlFetchedAtRef,
+        lastAudioRecoverySourceRef,
+    });
 
     usePlaybackUiEffects({
         statusMsg,
@@ -1528,156 +1025,6 @@ export default function App() {
         volumePreviewFrameRef,
         onClearPendingUnavailableSkip: clearPendingUnavailableSkip,
     });
-
-    const getPlayableOnlineQueue = useCallback((queue: SongResult[]) => {
-        return queue.filter(queuedSong => {
-            if (isLocalPlaybackSong(queuedSong) || isNavidromePlaybackSong(queuedSong)) {
-                return true;
-            }
-            return !isSongMarkedUnavailable(queuedSong);
-        });
-    }, []);
-
-    const getNextPlayableQueueSong = useCallback((queue: SongResult[], songId: number) => {
-        const currentIndex = queue.findIndex(queuedSong => queuedSong.id === songId);
-        if (currentIndex === -1) {
-            return null;
-        }
-
-        for (let index = currentIndex + 1; index < queue.length; index += 1) {
-            const candidate = queue[index];
-            if (isLocalPlaybackSong(candidate) || isNavidromePlaybackSong(candidate) || !isSongMarkedUnavailable(candidate)) {
-                return candidate;
-            }
-        }
-
-        if (loopMode === 'all' && queue.length > 1) {
-            for (let index = 0; index < currentIndex; index += 1) {
-                const candidate = queue[index];
-                if (isLocalPlaybackSong(candidate) || isNavidromePlaybackSong(candidate) || !isSongMarkedUnavailable(candidate)) {
-                    return candidate;
-                }
-            }
-        }
-
-        return null;
-    }, [loopMode]);
-
-    const buildQueueWithReplacementSong = useCallback((
-        queue: SongResult[],
-        originalSong: SongResult,
-        replacementSong: SongResult
-    ) => {
-        const normalizedQueue = queue.length > 0 ? queue : [originalSong];
-        const replacedQueue = normalizedQueue.flatMap((queuedSong) => {
-            if (isLocalPlaybackSong(queuedSong) || isNavidromePlaybackSong(queuedSong)) {
-                return [queuedSong];
-            }
-
-            if (queuedSong.id === originalSong.id) {
-                return [replacementSong];
-            }
-
-            if (isSongMarkedUnavailable(queuedSong)) {
-                return [];
-            }
-
-            return [queuedSong];
-        });
-
-        if (replacedQueue.length === 0) {
-            return [replacementSong];
-        }
-
-        if (!replacedQueue.some((queuedSong) => queuedSong.id === replacementSong.id)) {
-            replacedQueue.push(replacementSong);
-        }
-
-        return replacedQueue;
-    }, []);
-
-    const handleMarkedUnavailableSong = useCallback(async (
-        song: SongResult,
-        queue: SongResult[],
-        isFmCall: boolean,
-        options: PlaybackNavigationOptions
-    ) => {
-        setIsLyricsLoading(false);
-        setStatusMsg({ type: 'info', text: t('status.loadingSong') });
-        try {
-            const replacement = await neteaseApi.getUnavailableSongReplacement(song);
-
-            if (!replacement || !replacement.replacementSong || isSongMarkedUnavailable(replacement.replacementSong)) {
-                setStatusMsg({ type: 'error', text: t('status.songUnavailable') });
-                return true;
-            }
-
-            setStatusMsg(null);
-            setPendingUnavailableReplacement({
-                originalSong: song,
-                replacementSong: replacement.replacementSong,
-                replacementSongId: replacement.replacementSongId,
-                typeDesc: replacement.typeDesc,
-                queue,
-                isFmCall,
-                options,
-            });
-            return true;
-        } catch (error) {
-            console.error('[App] Failed to load replacement song before dialog:', error);
-            setStatusMsg({ type: 'error', text: t('status.playbackError') });
-            return true;
-        }
-    }, [t]);
-
-    const showTimedSkipPrompt = useCallback((
-        messageKey: SkipPromptMessageKey,
-        onSkip: () => void,
-        onCancel?: () => void
-    ) => {
-        clearPendingUnavailableSkip();
-
-        let remainingSeconds = Math.ceil(UNAVAILABLE_SKIP_CONFIRM_TIMEOUT_MS / 1000);
-        const skip = () => {
-            clearPendingUnavailableSkip();
-            setStatusMsg(null);
-            onSkip();
-        };
-        const cancel = () => {
-            clearPendingUnavailableSkip();
-            setStatusMsg(null);
-            onCancel?.();
-        };
-        const buildMessage = (seconds: number): StatusMessage => ({
-            type: 'error',
-            text: t(messageKey, { seconds }),
-            persistent: true,
-            actionLabel: t('status.skipUnavailableAction'),
-            cancelLabel: t('status.cancel'),
-            onAction: skip,
-            onCancel: cancel,
-        });
-
-        setStatusMsg(buildMessage(remainingSeconds));
-        pendingUnavailableSkipTimerRef.current = window.setTimeout(skip, UNAVAILABLE_SKIP_CONFIRM_TIMEOUT_MS);
-        pendingUnavailableSkipIntervalRef.current = window.setInterval(() => {
-            remainingSeconds -= 1;
-            if (remainingSeconds <= 0) {
-                if (pendingUnavailableSkipIntervalRef.current !== null) {
-                    window.clearInterval(pendingUnavailableSkipIntervalRef.current);
-                    pendingUnavailableSkipIntervalRef.current = null;
-                }
-                return;
-            }
-
-            setStatusMsg(current => {
-                if (!current?.persistent) {
-                    return current;
-                }
-                return buildMessage(remainingSeconds);
-            });
-        }, UNAVAILABLE_SKIP_CONFIRM_INTERVAL_MS);
-    }, [clearPendingUnavailableSkip, t]);
 
     // Audio Analyzer Setup
     // TODO: LOW PRIORITY::
@@ -1710,341 +1057,6 @@ export default function App() {
             console.error("Audio Context Setup Failed:", e);
         }
     };
-
-    const playSong = async (
-        song: SongResult,
-        queue: SongResult[] = [],
-        isFmCall: boolean = false,
-        options: PlaybackNavigationOptions = {}
-    ) => {
-        const restoredMainSnapshot = interruptStagePlaybackForMainTransition();
-
-        console.log("[App] playSong initiated:", song.name, song.id, "isFm:", isFmCall);
-        clearPendingUnavailableSkip();
-        setStatusMsg(prev => prev?.persistent ? null : prev);
-        const shouldNavigateToPlayer = options.shouldNavigateToPlayer ?? true;
-        setIsFmMode(isFmCall);
-        if (isFmCall && !isFmMode) {
-            // Only auto-open panel when first entering FM mode
-            setPanelTab('queue');
-            setIsPanelOpen(true);
-        }
-
-        const playbackRequestId = ++playbackRequestIdRef.current;
-        const isLatestPlaybackRequest = () => playbackRequestIdRef.current === playbackRequestId;
-        const isLocal = isLocalPlaybackSong(song);
-        const isNavidrome = isNavidromePlaybackSong(song);
-        let prefetched: ReturnType<typeof getPrefetchedData> = null;
-        let preloadedOnlineAudioResult: Awaited<ReturnType<typeof loadOnlineSongAudioSource>> | null = null;
-        const mainQueueContext = restoredMainSnapshot?.playQueue ?? playQueue;
-        const queueContext = queue.length > 0 ? queue : mainQueueContext.length === 0 ? [song] : mainQueueContext;
-        let newQueue = getPlayableOnlineQueue(queueContext);
-        const skipCount = options.unavailableSkipCount ?? 0;
-        playbackAutoSkipCountRef.current = skipCount;
-
-        if (!isLocal && !isNavidrome && isSongMarkedUnavailable(song)) {
-            if (await handleMarkedUnavailableSong(song, queueContext, isFmCall, options)) {
-                return;
-            }
-        }
-
-        if (!isLocal && !isNavidrome) {
-            prefetched = getPrefetchedData(song, audioQuality);
-
-            const hasImmediatePrefetchedAudio = Boolean(
-                prefetched?.audioUrl &&
-                prefetched.audioUrl !== 'CACHED_IN_DB'
-            );
-            const hasCachedAudioBlob = hasImmediatePrefetchedAudio
-                ? null
-                : await hasCachedAudio(getOnlineSongCacheKey('audio', song));
-
-            if (!isLatestPlaybackRequest()) return;
-
-            if (!hasImmediatePrefetchedAudio && !hasCachedAudioBlob) {
-                setStatusMsg({ type: 'info', text: t('status.loadingSong') });
-            }
-
-            try {
-                preloadedOnlineAudioResult = await loadOnlineSongAudioSource(song, audioQuality, prefetched);
-                if (!isLatestPlaybackRequest()) {
-                    if (preloadedOnlineAudioResult.kind === 'ok' && preloadedOnlineAudioResult.blobUrl) {
-                        URL.revokeObjectURL(preloadedOnlineAudioResult.blobUrl);
-                    }
-                    return;
-                }
-
-                if (preloadedOnlineAudioResult.kind === 'unavailable') {
-                    console.warn("[App] Song URL is empty, likely unavailable");
-
-                    const nextSong = getNextPlayableQueueSong(queueContext, song.id);
-                    const canSkip = Boolean(nextSong) && skipCount < MAX_UNAVAILABLE_AUTO_SKIP_COUNT;
-
-                    setIsLyricsLoading(false);
-
-                    if (canSkip && nextSong) {
-                        showTimedSkipPrompt('status.songUnavailablePrompt', () => {
-                            if (playbackRequestIdRef.current !== playbackRequestId) return;
-                            void playSong(nextSong, newQueue, isFmCall, {
-                                ...options,
-                                unavailableSkipCount: skipCount + 1
-                            });
-                        });
-                    } else {
-                        setStatusMsg({ type: 'error', text: t('status.songUnavailable') });
-                    }
-                    return;
-                }
-            } catch (e) {
-                console.error("[App] Failed to fetch song URL:", e);
-                setStatusMsg({ type: 'error', text: t('status.playbackError') });
-                setIsLyricsLoading(false);
-                return;
-            }
-        }
-
-        // Enable autoplay for user-initiated song changes
-        shouldAutoPlay.current = true;
-        currentSongRef.current = song.id;
-        pendingResumeTimeRef.current = null;
-        lastAudioRecoverySourceRef.current = null;
-        currentOnlineAudioUrlFetchedAtRef.current = null;
-
-        // 0. Instant UI Feedback
-        setLyrics(null);
-        setCurrentLineIndex(-1);
-        currentTime.set(0); // Reset currentTime to prevent stale playback position
-        setDuration(0);
-        setCurrentSong(song);
-        setCachedCoverUrl(null);
-        setAudioSrc(null);
-        setIsLyricsLoading(true); // Start loading lyrics
-
-        // Revoke old blob
-        if (blobUrlRef.current) {
-            URL.revokeObjectURL(blobUrlRef.current);
-            blobUrlRef.current = null;
-        }
-
-        // 1. Queue Management
-        if (queue.length > 0) {
-            setPlayQueue(newQueue);
-        } else if (playQueue.length === 0) {
-            setPlayQueue(newQueue);
-        }
-
-        // Save for next reload
-        void persistLastPlaybackCache(song, newQueue);
-
-        if (shouldNavigateToPlayer) {
-            navigateToPlayer();
-        }
-        setPlayerState(PlayerState.IDLE);
-
-        if (isNavidrome) {
-            const navidromeSong = resolveNavidromePlaybackCarrier(song);
-            if (!navidromeSong) {
-                setStatusMsg({ type: 'error', text: t('status.playbackError') });
-                setIsLyricsLoading(false);
-                return;
-            }
-
-            const navidromeQueue = queue.length > 0
-                ? queue
-                    .map(queuedSong => resolveNavidromePlaybackCarrier(queuedSong))
-                    .filter((queuedSong): queuedSong is NavidromeSong => Boolean(queuedSong))
-                : playQueue
-                    .map(queuedSong => resolveNavidromePlaybackCarrier(queuedSong))
-                    .filter((queuedSong): queuedSong is NavidromeSong => Boolean(queuedSong));
-
-            await onPlayNavidromeSong(navidromeSong, navidromeQueue, { shouldNavigateToPlayer: false });
-            return;
-        }
-
-        if (isLocal) {
-            console.log("[App] Playing Local Song");
-
-            // 2. Load Local Audio
-            let blobUrl: string | null = null;
-            let currentLocalData = (song as any).localData;
-
-            // Try to get blob from localData if available
-            if (currentLocalData) {
-                blobUrl = await getAudioFromLocalSong(currentLocalData);
-            }
-            // If not in song object (e.g. from queue restoration), try to find in localSongs list
-            else {
-                // Try to match by name/duration as fallback
-                const found = localSongs.find(ls =>
-                    (ls.title || ls.fileName) === song.name &&
-                    Math.abs(ls.duration - song.duration) < 1000
-                );
-                if (found) {
-                    currentLocalData = found;
-                    blobUrl = await getAudioFromLocalSong(found);
-                }
-            }
-
-            if (blobUrl) {
-                blobUrlRef.current = blobUrl;
-                setAudioSrc(blobUrl);
-
-                // 3. Instant Local Metadata + Background Auto-Match
-                if (currentLocalData) {
-                    currentLocalData = await ensureLocalSongEmbeddedCover(currentLocalData);
-                    const initialMeta = await resolveLocalMetadataUI(currentLocalData, null);
-                    setCurrentSong(initialMeta.unifiedSong);
-                    const localQueueContext = playQueue
-                        .map(queuedSong => (queuedSong as any).localData as LocalSong | undefined)
-                        .filter((queuedSong): queuedSong is LocalSong => Boolean(queuedSong));
-                    prewarmNearbyLocalSongs(currentLocalData, localQueueContext);
-                    
-                    if (initialMeta.coverUrl) {
-                        loadCachedOrFetchCover(`cover_local_${currentLocalData.id}`, initialMeta.coverUrl).then(res => {
-                            if (currentSongRef.current === song.id) setCachedCoverUrl(res);
-                        });
-                    } else {
-                        setCachedCoverUrl(null);
-                    }
-                    
-                    setLyrics(initialMeta.lyrics);
-                    
-                    const needsLyricsMatch = !currentLocalData.hasLocalLyrics && !currentLocalData.hasEmbeddedLyrics && !currentLocalData.matchedLyrics;
-                    const needsCoverMatch = !currentLocalData.embeddedCover && !currentLocalData.matchedCoverUrl;
-                    if ((needsLyricsMatch || needsCoverMatch) && !currentLocalData.noAutoMatch) {
-                        setIsLyricsLoading(true);
-                    } else {
-                        setIsLyricsLoading(false);
-                    }
-
-                    // Background match
-                    handleLocalSongMatch(currentLocalData).then(async ({ updatedLocalSong, matchedSongResult }) => {
-                        if (currentSongRef.current !== song.id) return;
-
-                        const updatedMeta = await resolveLocalMetadataUI(updatedLocalSong, matchedSongResult);
-                        
-                        setCurrentSong(updatedMeta.unifiedSong);
-                        setLyrics(updatedMeta.lyrics);
-                        setIsLyricsLoading(false);
-
-                        if (updatedMeta.coverUrl && updatedMeta.coverUrl !== initialMeta.coverUrl) {
-                            loadCachedOrFetchCover(`cover_local_${updatedLocalSong.id}`, updatedMeta.coverUrl).then(res => {
-                                if (currentSongRef.current === updatedMeta.unifiedSong.id) setCachedCoverUrl(res);
-                            });
-                        } else if (!updatedMeta.coverUrl) {
-                            if (currentSongRef.current === updatedMeta.unifiedSong.id) setCachedCoverUrl(null);
-                        }
-                    });
-                } else {
-                    setIsLyricsLoading(false);
-                }
-
-                // Theme
-                try {
-                    await restoreCachedThemeForSong(song.id);
-                    if (currentSongRef.current !== song.id) return;
-                } catch (e) {
-                    console.warn("Theme load error", e);
-                }
-
-                return; // EXIT for local songs
-            } else {
-                setStatusMsg({ type: 'error', text: '无法播放本地文件 (文件可能已移动或权限丢失)' });
-                setIsLyricsLoading(false);
-                return;
-            }
-        }
-
-        // --- ONLINE SONG LOGIC BELOW ---
-
-        // 2. Load Cached Cover (Visual Feedback)
-        const cachedCoverUrl = await getCachedCoverUrl(getOnlineSongCacheKey('cover', song));
-        if (currentSongRef.current !== song.id) return;
-        if (cachedCoverUrl) {
-            setCachedCoverUrl(cachedCoverUrl);
-        } else if (prefetched?.coverUrl) {
-            // Use prefetched cover URL as fallback
-            setCachedCoverUrl(prefetched.coverUrl);
-        }
-
-        // 3. Audio Loading (Prefetch Cache vs IndexedDB vs Network)
-        const audioResult = preloadedOnlineAudioResult;
-        if (!audioResult || audioResult.kind !== 'ok') {
-            setStatusMsg({ type: 'error', text: t('status.playbackError') });
-            setPlayerState(PlayerState.IDLE);
-            setIsLyricsLoading(false);
-            return;
-        }
-
-        if (audioResult.blobUrl) {
-            blobUrlRef.current = audioResult.blobUrl;
-            currentOnlineAudioUrlFetchedAtRef.current = null;
-        } else if (audioResult.audioSrc.startsWith('http')) {
-            currentOnlineAudioUrlFetchedAtRef.current =
-                prefetched?.audioUrl === audioResult.audioSrc
-                    ? prefetched.audioUrlFetchedAt
-                    : Date.now();
-        } else {
-            currentOnlineAudioUrlFetchedAtRef.current = null;
-        }
-        setAudioSrc(audioResult.audioSrc);
-
-        // 4. Fetch Lyrics (Prefetch Cache vs IndexedDB vs Network)
-        try {
-            await loadOnlineSongLyrics(song, prefetched, user?.userId, {
-                isCurrent: () => currentSongRef.current === song.id,
-                onLyrics: (resolvedLyrics) => setLyrics(resolvedLyrics),
-                onPureMusicChange: (isPureMusic) => {
-                    setCurrentSong(prev => prev?.id === song.id ? { ...prev, isPureMusic } : prev);
-                },
-                onDone: () => setIsLyricsLoading(false),
-            });
-        } catch (e) {
-            console.warn("[App] Lyric fetch failed", e);
-            setLyrics(null);
-            setIsLyricsLoading(false); // Failed
-        }
-
-        // 5. Handle Theme
-        try {
-            await restoreCachedThemeForSong(song.id);
-            if (currentSongRef.current !== song.id) return;
-        } catch (e) {
-            console.warn("Theme load error", e);
-        }
-
-        // 6. Trigger prefetch for nearby songs in queue
-        if (newQueue.length > 1) {
-            prefetchNearbySongs(song.id, newQueue, audioQuality, user?.userId);
-        }
-    };
-
-    const handleStageExternalPlayRequest = useCallback(async (request: { requestId: string; songId: number; appendToQueue?: boolean; }) => {
-        try {
-            const detail = await neteaseApi.getSongDetail(request.songId);
-            const song = (detail?.songs || [])[0] as SongResult | undefined;
-            if (!song) {
-                throw new Error(`Song ${request.songId} was not found.`);
-            }
-
-            if (request.appendToQueue) {
-                appendNeteaseSongsToMainQueue([song]);
-            } else {
-                await playSong(song, [song], false, { shouldNavigateToPlayer: true });
-            }
-            await window.electron?.completeStageExternalPlayRequest?.({
-                requestId: request.requestId,
-                ok: true,
-            });
-        } catch (error) {
-            console.warn('[Stage] Failed to handle external play request', error);
-            await window.electron?.completeStageExternalPlayRequest?.({
-                requestId: request.requestId,
-                ok: false,
-                error: error instanceof Error ? error.message : String(error),
-            });
-        }
-    }, [appendNeteaseSongsToMainQueue, playSong]);
 
     const cacheSongAssets = async () => {
         if (!currentSong || !audioSrc || audioSrc.startsWith('blob:')) return;
@@ -2085,228 +1097,6 @@ export default function App() {
         // Update usage if panel is open
         if (isPanelOpen && panelTab === 'account') updateCacheSize();
     };
-
-    const playOnlineQueueFromStart = (songs: SongResult[]) => {
-        const playableSongs = getPlayableOnlineQueue(songs);
-        if (playableSongs.length === 0) {
-            setStatusMsg({ type: 'error', text: t('status.noPlayableSongs') });
-            return;
-        }
-
-        void playSong(playableSongs[0], playableSongs, false);
-    };
-
-    const handleQueueAddAndPlay = (song: SongResult) => {
-        // Check if song exists in queue
-        const existingIndex = playQueue.findIndex(s => s.id === song.id);
-
-        let newQueue = [...playQueue];
-
-        if (existingIndex === -1) {
-            // Add to end if not present (or after current song? Let's just append)
-            newQueue.push(song);
-        }
-
-        // Play this song, updating the queue state
-        playSong(song, newQueue, false);
-    };
-
-    const handleSearchOverlaySubmit = useCallback(async () => {
-        const trimmedQuery = searchQuery.trim();
-        if (!trimmedQuery) {
-            return;
-        }
-
-        const didSearch = await submitSearch({
-            query: trimmedQuery,
-            sourceTab: searchSourceTab,
-            deps: {
-                localSongs,
-                t: (key, fallback) => t(key, fallback ?? ''),
-            },
-        });
-
-        if (didSearch) {
-            navigateToSearch({
-                query: trimmedQuery,
-                sourceTab: searchSourceTab,
-                replace: Boolean(window.history.state?.search),
-            });
-        }
-    }, [localSongs, navigateToSearch, searchQuery, searchSourceTab, submitSearch, t]);
-
-    const handleSearchLoadMore = useCallback(async () => {
-        await loadMoreSearchResults({
-            deps: {
-                localSongs,
-                t: (key, fallback) => t(key, fallback ?? ''),
-            },
-        });
-    }, [loadMoreSearchResults, localSongs, t]);
-
-    const handleSearchResultPlay = useCallback((track: UnifiedSong) => {
-        if (track.isLocal && track.localData) {
-            void onPlayLocalSong(track.localData);
-            return;
-        }
-
-        if (track.isNavidrome && track.navidromeData) {
-            void onPlayNavidromeSong(track as NavidromeSong);
-            return;
-        }
-
-        handleQueueAddAndPlay(track);
-    }, [handleQueueAddAndPlay, onPlayLocalSong, onPlayNavidromeSong]);
-
-    const handleUnavailableReplacementConfirm = useCallback(async () => {
-        if (!pendingUnavailableReplacement) {
-            return;
-        }
-
-        const { originalSong, replacementSong, replacementSongId, queue, isFmCall, options } = pendingUnavailableReplacement;
-        setPendingUnavailableReplacement(null);
-
-        try {
-            if (!replacementSong || replacementSong.id !== replacementSongId || isSongMarkedUnavailable(replacementSong)) {
-                setStatusMsg({ type: 'error', text: t('status.songUnavailable') });
-                return;
-            }
-
-            const replacementQueue = buildQueueWithReplacementSong(queue, originalSong, replacementSong);
-            await playSong(replacementSong, replacementQueue, isFmCall, options);
-        } catch (error) {
-            console.error('[App] Failed to load replacement song:', error);
-            setStatusMsg({ type: 'error', text: t('status.playbackError') });
-        }
-    }, [buildQueueWithReplacementSong, pendingUnavailableReplacement, t]);
-
-    const handleSearchResultArtistSelect = useCallback((track: UnifiedSong, artistName: string, artistId?: number) => {
-        if (track.isLocal) {
-            hideSearchOverlay();
-            openLocalArtistByName(artistName);
-            return;
-        }
-
-        if (track.isNavidrome && track.navidromeData?.artistId) {
-            hideSearchOverlay();
-            setHomeViewTab('navidrome');
-            setPendingNavidromeSelection({ type: 'artist', artistId: track.navidromeData.artistId });
-            return;
-        }
-
-        if (artistId) {
-            handleArtistSelect(artistId);
-        }
-    }, [handleArtistSelect, hideSearchOverlay, openLocalArtistByName, setHomeViewTab]);
-
-    const handleSearchResultAlbumSelect = useCallback((track: UnifiedSong, albumName: string, albumId?: number) => {
-        if (track.isLocal) {
-            hideSearchOverlay();
-            openLocalAlbumByName(albumName);
-            return;
-        }
-
-        if (track.isNavidrome && track.navidromeData?.albumId) {
-            hideSearchOverlay();
-            setHomeViewTab('navidrome');
-            setPendingNavidromeSelection({ type: 'album', albumId: track.navidromeData.albumId });
-            return;
-        }
-
-        if (albumId) {
-            handleAlbumSelect(albumId);
-        }
-    }, [handleAlbumSelect, hideSearchOverlay, openLocalAlbumByName, setHomeViewTab]);
-
-    const handleNextTrack = useCallback(async (options?: NextTrackOptions) => {
-        if (isNowPlayingStageActive) return;
-        if (!currentSong || playQueue.length === 0) return;
-
-        const shouldNavigateToPlayer = options?.shouldNavigateToPlayer ?? true;
-        const currentIndex = playQueue.findIndex(s => s.id === currentSong.id);
-
-        // --- FM Mode Auto-fetch ---
-        if (isFmMode && currentIndex >= playQueue.length - 2) {
-            try {
-                const fmRes = await neteaseApi.getPersonalFm();
-                if (fmRes.data && fmRes.data.length > 0) {
-                    const newQueue = [...playQueue, ...fmRes.data];
-                    setPlayQueue(newQueue);
-                    playSong(newQueue[currentIndex + 1], newQueue, true, {
-                        shouldNavigateToPlayer,
-                        unavailableSkipCount: options?.unavailableSkipCount
-                    });
-                    return;
-                }
-            } catch (e) {
-                console.error("Failed to fetch FM tracks", e);
-            }
-        }
-
-        let nextIndex = -1;
-
-        if (currentIndex >= 0 && currentIndex < playQueue.length - 1) {
-            nextIndex = currentIndex + 1;
-        } else if (loopMode === 'all') {
-            // Wrap around
-            nextIndex = 0;
-        }
-
-        if (nextIndex >= 0) {
-            playSong(playQueue[nextIndex], playQueue, isFmMode, {
-                shouldNavigateToPlayer,
-                unavailableSkipCount: options?.unavailableSkipCount
-            });
-        } else if (options?.allowStopOnMissing) {
-            if (audioRef.current) {
-                audioRef.current.pause();
-            }
-            setPlayerState(PlayerState.IDLE);
-        }
-    }, [currentSong, isFmMode, isNowPlayingStageActive, loopMode, playQueue]);
-
-    const handlePrevTrack = useCallback(() => {
-        if (isNowPlayingStageActive) return;
-        if (!currentSong || playQueue.length === 0) return;
-
-        const currentIndex = playQueue.findIndex(s => s.id === currentSong.id);
-        let prevIndex = -1;
-
-        if (currentIndex > 0) {
-            prevIndex = currentIndex - 1;
-        } else if (loopMode === 'all') {
-            prevIndex = playQueue.length - 1;
-        }
-
-        if (prevIndex >= 0) {
-            playSong(playQueue[prevIndex], playQueue, isFmMode);
-        }
-    }, [currentSong, isFmMode, isNowPlayingStageActive, loopMode, playQueue]);
-
-    const skipAfterPlaybackFailure = useCallback(() => {
-        clearPendingUnavailableSkip();
-        const skipCount = playbackAutoSkipCountRef.current;
-        const currentIndex = currentSong ? playQueue.findIndex(song => song.id === currentSong.id) : -1;
-        const hasNextTrack = currentIndex >= 0 && (
-            currentIndex < playQueue.length - 1 ||
-            (loopMode === 'all' && playQueue.length > 1)
-        );
-
-        if (!hasNextTrack || skipCount >= MAX_UNAVAILABLE_AUTO_SKIP_COUNT) {
-            setPlayerState(PlayerState.IDLE);
-            return;
-        }
-
-        const nextSkipCount = skipCount + 1;
-        showTimedSkipPrompt('status.playbackErrorPrompt', () => {
-            playbackAutoSkipCountRef.current = nextSkipCount;
-            void handleNextTrack({
-                allowStopOnMissing: true,
-                shouldNavigateToPlayer: false,
-                unavailableSkipCount: nextSkipCount
-            });
-        });
-    }, [clearPendingUnavailableSkip, currentSong, handleNextTrack, loopMode, playQueue, showTimedSkipPrompt]);
 
     const mediaSessionPlayRef = useRef(resumePlayback);
     const mediaSessionPauseRef = useRef(pausePlayback);
@@ -2386,42 +1176,6 @@ export default function App() {
             handleNextTrack();
         }
     };
-
-    const shuffleQueue = useCallback(() => {
-        if (isNowPlayingStageActive) return;
-        if (!playQueue || playQueue.length <= 1) return;
-
-        // 1. Identify current song
-        const currentId = currentSong?.id;
-
-        // 2. Separate current song from others
-        let songsToShuffle: SongResult[] = [];
-        let firstSong: SongResult | null = null;
-
-        if (currentId) {
-            firstSong = playQueue.find(s => s.id === currentId) || null;
-            songsToShuffle = playQueue.filter(s => s.id !== currentId);
-        } else {
-            songsToShuffle = [...playQueue];
-        }
-
-        // 3. Fisher-Yates Shuffle
-        for (let i = songsToShuffle.length - 1; i > 0; i--) {
-            const j = Math.floor(Math.random() * (i + 1));
-            [songsToShuffle[i], songsToShuffle[j]] = [songsToShuffle[j], songsToShuffle[i]];
-        }
-
-        // 4. Reconstruct queue: Current song first (if exists), then shuffled songs
-        const newQueue = firstSong ? [firstSong, ...songsToShuffle] : songsToShuffle;
-
-        setPlayQueue(newQueue);
-        setStatusMsg({ type: 'success', text: t('status.queueShuffled') || 'Queue Shuffled' });
-
-        // 5. Re-prefetch based on new queue order
-        if (currentId && newQueue.length > 1) {
-            invalidateAndRefetch(currentId, newQueue, audioQuality);
-        }
-    }, [audioQuality, currentSong, isNowPlayingStageActive, playQueue, t]);
 
     // Volume & Mute Sync
     useEffect(() => {
@@ -2828,228 +1582,14 @@ export default function App() {
         handleToggleLoopMode();
     };
 
-    const handleLike = async () => {
-        if (!currentSong) return;
-
-        if (isStagePlaybackSong(currentSong)) {
-            setStatusMsg({ type: 'info', text: t('status.stageActionUnavailable') || 'Stage 模式下不支持收藏操作' });
-            return;
-        }
-
-        if (isLocalPlaybackSong(currentSong) && currentSong.localData) {
-            const nextLiked = !isLocalSongLiked(currentSong);
-            try {
-                await setLocalSongFavorite(currentSong.localData, nextLiked);
-                await loadLocalPlaylists();
-                setStatusMsg({ type: 'success', text: nextLiked ? t('status.liked') : (t('status.unliked') || '已取消喜欢') });
-            } catch (e) {
-                console.error('Failed to update local favorite playlist', e);
-                setStatusMsg({ type: 'error', text: t('status.likeFailed') });
-            }
-            return;
-        }
-
-        if (isNavidromePlaybackSong(currentSong)) {
-            setStatusMsg({ type: 'info', text: 'Navidrome 歌曲的收藏能力尚未接入' });
-            return;
-        }
-
-        const isLiked = likedSongIds.has(currentSong.id);
-        const newStatus = !isLiked;
-
-        try {
-            await neteaseApi.likeSong(currentSong.id, newStatus);
-
-            // Update local state immediately
-            setLikedSongIds(prev => {
-                const next = new Set(prev);
-                if (newStatus) next.add(currentSong.id);
-                else next.delete(currentSong.id);
-
-                // Update cache
-                saveToCache('user_liked_songs', Array.from(next));
-
-                return next;
-            });
-
-            setStatusMsg({ type: 'success', text: newStatus ? t('status.liked') : t('status.unliked') || "Removed from Liked" });
-        } catch (e) {
-            console.error("Like failed", e);
-            setStatusMsg({ type: 'error', text: t('status.likeFailed') });
-        }
-    };
-
-    const handleUpdateLocalLyrics = async (content: string, isTranslation: boolean) => {
-        if (!isLocalPlaybackSong(currentSong)) return;
-
-        const localData = (currentSong as any).localData as LocalSong;
-        if (!localData) return;
-
-        const updatedLocalSong = { ...localData };
-        if (isTranslation) {
-            updatedLocalSong.hasLocalTranslationLyrics = true;
-            updatedLocalSong.localTranslationLyricsContent = content;
-        } else {
-            updatedLocalSong.hasLocalLyrics = true;
-            updatedLocalSong.localLyricsContent = content;
-        }
-
-        // Save to DB
-        try {
-            const { saveLocalSong } = await import('./services/db');
-            await saveLocalSong(updatedLocalSong);
-
-            // Re-run onPlayLocalSong to handle everything (parsing, merging) properly
-            onPlayLocalSong(updatedLocalSong, localSongs);
-
-            setStatusMsg({ type: 'success', text: isTranslation ? 'Translation lyrics updated' : 'Lyrics updated' });
-        } catch (e) {
-            console.error("Failed to save local lyrics", e);
-            setStatusMsg({ type: 'error', text: 'Failed to save lyrics' });
-        }
-    };
-
-    const handleChangeLyricsSource = async (source: 'local' | 'embedded' | 'online') => {
-        if (!isLocalPlaybackSong(currentSong)) return;
-
-        const localData = (currentSong as any).localData as LocalSong;
-        if (!localData) return;
-
-        const updatedLocalSong = { ...localData };
-        updatedLocalSong.lyricsSource = source;
-
-        // Save to DB
-        try {
-            const { saveLocalSong } = await import('./services/db');
-            await saveLocalSong(updatedLocalSong);
-
-            // Parse and apply lyrics from the selected source
-            let newLyrics: LyricData | null = null;
-            if (source === 'local' && updatedLocalSong.localLyricsContent) {
-                newLyrics = await LyricParserFactory.parse({ type: 'local', lrcContent: updatedLocalSong.localLyricsContent, tLrcContent: updatedLocalSong.localTranslationLyricsContent });
-            } else if (source === 'embedded' && updatedLocalSong.embeddedLyricsContent) {
-                newLyrics = await LyricParserFactory.parse({ type: 'embedded', textContent: updatedLocalSong.embeddedLyricsContent, translationContent: updatedLocalSong.embeddedTranslationLyricsContent });
-            } else if (source === 'online' && updatedLocalSong.matchedLyrics) {
-                newLyrics = updatedLocalSong.matchedLyrics;
-            }
-            setLyrics(newLyrics);
-            setCurrentLineIndex(-1);
-
-            // Update current song's localData reference
-            const updatedSong = { ...currentSong };
-            (updatedSong as any).localData = updatedLocalSong;
-            setCurrentSong(updatedSong);
-
-            // Refresh local songs list
-            await loadLocalSongs();
-
-            setStatusMsg({ type: 'success', text: '歌词来源已切换' });
-        } catch (e) {
-            console.error("Failed to save lyrics source", e);
-            setStatusMsg({ type: 'error', text: 'Failed to save lyrics source' });
-        }
-    };
-
     const handleChangeReplayGainMode = (mode: ReplayGainMode) => {
         setReplayGainMode(mode);
         setStatusMsg({ type: 'info', text: replayGainModeLabels[mode] });
     };
 
-    const [showLyricMatchModal, setShowLyricMatchModal] = useState(false);
-    const [showNaviLyricMatchModal, setShowNaviLyricMatchModal] = useState(false);
-
-    const handleManualMatchOnline = () => {
-        setIsPanelOpen(false);
-        if (currentSong && (currentSong as any).isNavidrome) {
-            setShowNaviLyricMatchModal(true);
-            return;
-        }
-
-        if (!isLocalPlaybackSong(currentSong)) return;
-        const localData = (currentSong as any).localData as LocalSong;
-        if (!localData) return;
-        setShowLyricMatchModal(true);
-    };
-
-    const handleLyricMatchComplete = async () => {
-        setShowLyricMatchModal(false);
-        if (!isLocalPlaybackSong(currentSong)) return;
-        const localData = (currentSong as any).localData as LocalSong;
-        if (!localData) return;
-
-        await loadLocalSongs();
-        const updatedList = await getLocalSongs();
-        const found = updatedList.find(s => s.id === localData.id);
-        if (found) {
-            onPlayLocalSong(found, localSongs);
-            setStatusMsg({ type: 'success', text: 'Match successful' });
-        }
-    };
-
-    const handleNaviLyricMatchComplete = async () => {
-        setShowNaviLyricMatchModal(false);
-        if (currentSong && (currentSong as any).isNavidrome) {
-            const navidromeQueue = playQueue
-                .map(song => (song as any).navidromeData as NavidromeSong | undefined)
-                .filter((song): song is NavidromeSong => Boolean(song?.isNavidrome));
-            onPlayNavidromeSong((currentSong as any).navidromeData, navidromeQueue);
-            setStatusMsg({ type: 'success', text: 'Match successful' });
-        }
-    };
-
-    const handleHomeMatchSong = useCallback(async (song: LocalSong) => {
-        await loadLocalSongs();
-
-        if (isLocalPlaybackSong(currentSong)) {
-            const currentLocalData = (currentSong as any).localData as LocalSong | undefined;
-            if (currentLocalData && currentLocalData.id === song.id) {
-                const updatedSongs = await getLocalSongs();
-                const updatedSong = updatedSongs.find(s => s.id === song.id);
-
-                if (updatedSong) {
-                    const updatedCurrentSong = { ...currentSong };
-                    (updatedCurrentSong as any).localData = updatedSong;
-
-                    if (updatedSong.matchedCoverUrl) {
-                        const coverUrl = updatedSong.matchedCoverUrl;
-                        if (updatedCurrentSong.al) {
-                            updatedCurrentSong.al.picUrl = coverUrl;
-                        } else {
-                            updatedCurrentSong.al = {
-                                id: 0,
-                                name: '',
-                                picUrl: coverUrl
-                            };
-                        }
-                    } else if (updatedCurrentSong.al) {
-                        updatedCurrentSong.al.picUrl = undefined;
-                    }
-
-                    setCurrentSong(updatedCurrentSong);
-
-                    if (updatedSong.matchedCoverUrl) {
-                        try {
-                            const response = await fetch(updatedSong.matchedCoverUrl, { mode: 'cors' });
-                            const coverBlob = await response.blob();
-                            await saveToCache(`cover_local_${updatedSong.id}`, coverBlob);
-                            setCachedCoverUrl(URL.createObjectURL(coverBlob));
-                        } catch (e) {
-                            console.warn('Failed to cache updated cover:', e);
-                            setCachedCoverUrl(updatedSong.matchedCoverUrl);
-                        }
-                    } else {
-                        setCachedCoverUrl(null);
-                    }
-
-                    if (updatedSong.matchedLyrics) {
-                        setLyrics(updatedSong.matchedLyrics);
-                    } else {
-                        setLyrics(null);
-                    }
-                }
-            }
-        }
-    }, [currentSong, loadLocalSongs, setLyrics]);
+    const handleLike = useCallback(async () => {
+        await handleLibraryLike(likedSongIds, setLikedSongIds);
+    }, [handleLibraryLike, likedSongIds, setLikedSongIds]);
 
     const handleContainerClick = () => {
         if (isPanelOpen) setIsPanelOpen(false);
